+++ src/reports/package/accessibilityInsightsReport.d.ts
