+++ src/tests/unit/tests/reports/package/reporter.test.ts
@@ -1,9 +1,9 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 import axe from 'axe-core';
-import AccessibilityInsightsReport from 'reports/library/accessibilityInsightsReport';
-import { AxeResultReport as AxeResultsReport } from 'reports/library/axe-results-report';
-import { AxeResultsReportGenerator, Reporter } from 'reports/library/reporter';
+import AccessibilityInsightsReport from 'reports/package/accessibilityInsightsReport';
+import { AxeResultReport as AxeResultsReport } from 'reports/package/axe-results-report';
+import { AxeResultsReportGenerator, Reporter } from 'reports/package/reporter';
 import { Mock } from 'typemoq';
 
 describe('Reporter', () => {
