+++ src/tests/unit/tests/reports/package/scans/scan-no-issues.ts
